modify: using common viewmodel to contain visibility state
