+++ app/src/main/java/com/qrist/quicker/qrlist/QRContainerFragment.kt
@@ -45,8 +45,14 @@ class QRContainerFragment : Fragment(), CoroutineScope {
     private lateinit var adapter: QRViewFragmentPagerAdapter
     private lateinit var sequence: TapTargetSequence
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
-        inflater.inflate(R.layout.fragment_qrcontainer, container, false)
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        viewModel.fetchServiceNameVisibility()
+        return inflater.inflate(R.layout.fragment_qrcontainer, container, false)
+    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
